% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDCN-package.R
\docType{package}
\name{BDCN-package}
\alias{BDCN}
\alias{BDCN-package}
\title{BDCN: Bayesian Dynamic Count Networks}
\description{
The package provides a reusable implementation of the graph-shrinkage BDCN
sampler, joint DSS, sequential prediction, and the sparse multi-hop
simulation used by the accompanying research project.
}
\author{
\strong{Maintainer}: Sheng Jiang \email{jiangsheng@cuhk.edu.cn}

Authors:
\itemize{
  \item Sheng Jiang \email{jiangsheng@cuhk.edu.cn}
  \item Ying Wu
}

}
\keyword{internal}
